Move instrument calibration functionality to be shared
