+++ api/src/opentrons/protocol_engine/resources/labware_data_provider.py
@@ -10,8 +10,7 @@ from typing import Optional
 from opentrons.protocols.models import LabwareDefinition
 from opentrons.protocols.labware import get_labware_definition
 
-# TODO (lc 09-26-2022) We should conditionally import ot2 or ot3 calibration
-from opentrons.hardware_control.instruments.ot2 import (
+from opentrons.hardware_control.instruments import (
     instrument_calibration as instr_cal,
 )
 from opentrons.calibration_storage.types import TipLengthCalNotFound
